debug seaming

+++ dr2/seaming.py
@@ -486,7 +486,7 @@ def run_all(lon1=25, lon2=215, ncores=2):
 """
 def seam(clusterview, lon1=25, lon2=215):
     """Warning: nodes must have ~5 GB memory each at l < 100"""
-    strips = np.arange(lon1, lon2+0.1, constants.STRIPWIDTH)
+    strips = np.arange(lon1, lon2+0.1, constants.STRIPWIDTH, dtype='int')
     log.info('Seaming in longitude strips %s' % (strips))
     clusterview.map(seam_one, strips, block=True)
